Add more environment variables.
Add a FAULTINJECT_LIBRARY_NAME environment variable, that causes the library
to ignore calls where it can't find the substring given in the (first ten levels)
stack strace. If it's not present all calls will be analyzed.

Add in a FAULTINJECT_XXX_COUNT environment variable for each function wrapper.

A couple of other bits of cleanup.